migrated to postgres, cards model added for payment integration
