+++ alembic.ini
@@ -63,6 +63,8 @@ version_path_separator = os
 # are written from script.py.mako
 # output_encoding = utf-8
 
+;sqlalchemy.url = driver://user:pass@localhost/dbname
+
 
 [post_write_hooks]
 # post_write_hooks defines scripts or Python functions that are run
