+++ alembic/env.py
@@ -11,6 +11,7 @@ from core.database import Base, DATABASE_URL
 from auth.models import *
 from user.models import *
 from visa_center.models import *
+from payment.models import *
 from order.models import *
 
 # this is the Alembic Config object, which provides
