--- alembic/versions/02ce91c4e140_initial_migration.py
@@ -1,123 +1,0 @@
-"""Initial migration
-
-Revision ID: 02ce91c4e140
-Revises: 09bdd6aa0849
-Create Date: 2025-01-26 13:53:46.630806
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '02ce91c4e140'
-down_revision: Union[str, None] = '09bdd6aa0849'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('phone_otps',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('phone_number', sa.String(), nullable=True),
-    sa.Column('otp', sa.String(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_phone_otps_id'), 'phone_otps', ['id'], unique=False)
-    op.create_index(op.f('ix_phone_otps_otp'), 'phone_otps', ['otp'], unique=False)
-    op.create_index(op.f('ix_phone_otps_phone_number'), 'phone_otps', ['phone_number'], unique=True)
-    op.create_table('users',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('username', sa.String(), nullable=True),
-    sa.Column('phone_number', sa.String(), nullable=True),
-    sa.Column('is_verified', sa.Boolean(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.Column('first_name', sa.String(), nullable=True),
-    sa.Column('last_name', sa.String(), nullable=True),
-    sa.Column('family_name', sa.String(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_users_family_name'), 'users', ['family_name'], unique=False)
-    op.create_index(op.f('ix_users_first_name'), 'users', ['first_name'], unique=False)
-    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
-    op.create_index(op.f('ix_users_last_name'), 'users', ['last_name'], unique=False)
-    op.create_index(op.f('ix_users_phone_number'), 'users', ['phone_number'], unique=True)
-    op.create_index(op.f('ix_users_username'), 'users', ['username'], unique=False)
-    op.create_table('emails',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('email', sa.String(), nullable=False),
-    sa.Column('is_verified', sa.Boolean(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_emails_email'), 'emails', ['email'], unique=True)
-    op.create_index(op.f('ix_emails_id'), 'emails', ['id'], unique=False)
-    op.create_index(op.f('ix_emails_user_id'), 'emails', ['user_id'], unique=True)
-    op.create_table('passports',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('passport_number', sa.String(), nullable=False),
-    sa.Column('passport_type', sa.Enum('ORDINARY', 'DIPLOMATIC', 'COLLECTIVE', 'SERVICE', 'OFFICIAL', 'FOREIGNERS', 'PROTECTION', 'UN', name='passporttype'), nullable=False),
-    sa.Column('issuer_country', sa.String(), nullable=False),
-    sa.Column('issue_date', sa.DateTime(), nullable=False),
-    sa.Column('expire_date', sa.DateTime(), nullable=False),
-    sa.Column('issue_place', sa.String(), nullable=False),
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_passports_id'), 'passports', ['id'], unique=False)
-    op.create_index(op.f('ix_passports_issue_place'), 'passports', ['issue_place'], unique=False)
-    op.create_index(op.f('ix_passports_issuer_country'), 'passports', ['issuer_country'], unique=False)
-    op.create_index(op.f('ix_passports_passport_number'), 'passports', ['passport_number'], unique=True)
-    op.create_index(op.f('ix_passports_passport_type'), 'passports', ['passport_type'], unique=False)
-    op.create_index(op.f('ix_passports_user_id'), 'passports', ['user_id'], unique=True)
-    op.create_table('tokens',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.Column('token', sa.String(), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_tokens_id'), 'tokens', ['id'], unique=False)
-    op.create_index(op.f('ix_tokens_token'), 'tokens', ['token'], unique=False)
-    op.create_index(op.f('ix_tokens_user_id'), 'tokens', ['user_id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_tokens_user_id'), table_name='tokens')
-    op.drop_index(op.f('ix_tokens_token'), table_name='tokens')
-    op.drop_index(op.f('ix_tokens_id'), table_name='tokens')
-    op.drop_table('tokens')
-    op.drop_index(op.f('ix_passports_user_id'), table_name='passports')
-    op.drop_index(op.f('ix_passports_passport_type'), table_name='passports')
-    op.drop_index(op.f('ix_passports_passport_number'), table_name='passports')
-    op.drop_index(op.f('ix_passports_issuer_country'), table_name='passports')
-    op.drop_index(op.f('ix_passports_issue_place'), table_name='passports')
-    op.drop_index(op.f('ix_passports_id'), table_name='passports')
-    op.drop_table('passports')
-    op.drop_index(op.f('ix_emails_user_id'), table_name='emails')
-    op.drop_index(op.f('ix_emails_id'), table_name='emails')
-    op.drop_index(op.f('ix_emails_email'), table_name='emails')
-    op.drop_table('emails')
-    op.drop_index(op.f('ix_users_username'), table_name='users')
-    op.drop_index(op.f('ix_users_phone_number'), table_name='users')
-    op.drop_index(op.f('ix_users_last_name'), table_name='users')
-    op.drop_index(op.f('ix_users_id'), table_name='users')
-    op.drop_index(op.f('ix_users_first_name'), table_name='users')
-    op.drop_index(op.f('ix_users_family_name'), table_name='users')
-    op.drop_table('users')
-    op.drop_index(op.f('ix_phone_otps_phone_number'), table_name='phone_otps')
-    op.drop_index(op.f('ix_phone_otps_otp'), table_name='phone_otps')
-    op.drop_index(op.f('ix_phone_otps_id'), table_name='phone_otps')
-    op.drop_table('phone_otps')
-    # ### end Alembic commands ###
