--- alembic/versions/09bdd6aa0849_initial_migration.py
@@ -1,30 +1,0 @@
-"""Initial migration
-
-Revision ID: 09bdd6aa0849
-Revises: 
-Create Date: 2025-01-26 13:51:34.130879
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '09bdd6aa0849'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
