--- alembic/versions/2b4bda5a4b63_passport_number_is_not_unique.py
@@ -1,32 +1,0 @@
-"""passport number is not unique
-
-Revision ID: 2b4bda5a4b63
-Revises: 849602ad0ca5
-Create Date: 2025-02-13 17:03:12.383892
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '2b4bda5a4b63'
-down_revision: Union[str, None] = '849602ad0ca5'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index('ix_passports_passport_number', table_name='passports')
-    op.create_index(op.f('ix_passports_passport_number'), 'passports', ['passport_number'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_passports_passport_number'), table_name='passports')
-    op.create_index('ix_passports_passport_number', 'passports', ['passport_number'], unique=1)
-    # ### end Alembic commands ###
