--- alembic/versions/2d8fc51d711f_passport_model_removed_from_user.py
@@ -1,53 +1,0 @@
-"""passport model removed from user
-
-Revision ID: 2d8fc51d711f
-Revises: 02ce91c4e140
-Create Date: 2025-01-27 21:19:05.224882
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '2d8fc51d711f'
-down_revision: Union[str, None] = '02ce91c4e140'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index('ix_passports_id', table_name='passports')
-    op.drop_index('ix_passports_issue_place', table_name='passports')
-    op.drop_index('ix_passports_issuer_country', table_name='passports')
-    op.drop_index('ix_passports_passport_number', table_name='passports')
-    op.drop_index('ix_passports_passport_type', table_name='passports')
-    op.drop_index('ix_passports_user_id', table_name='passports')
-    op.drop_table('passports')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('passports',
-    sa.Column('id', sa.INTEGER(), nullable=False),
-    sa.Column('passport_number', sa.VARCHAR(), nullable=False),
-    sa.Column('passport_type', sa.VARCHAR(length=10), nullable=False),
-    sa.Column('issuer_country', sa.VARCHAR(), nullable=False),
-    sa.Column('issue_date', sa.DATETIME(), nullable=False),
-    sa.Column('expire_date', sa.DATETIME(), nullable=False),
-    sa.Column('issue_place', sa.VARCHAR(), nullable=False),
-    sa.Column('user_id', sa.INTEGER(), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index('ix_passports_user_id', 'passports', ['user_id'], unique=1)
-    op.create_index('ix_passports_passport_type', 'passports', ['passport_type'], unique=False)
-    op.create_index('ix_passports_passport_number', 'passports', ['passport_number'], unique=1)
-    op.create_index('ix_passports_issuer_country', 'passports', ['issuer_country'], unique=False)
-    op.create_index('ix_passports_issue_place', 'passports', ['issue_place'], unique=False)
-    op.create_index('ix_passports_id', 'passports', ['id'], unique=False)
-    # ### end Alembic commands ###
