--- alembic/versions/31be58f3cea3_visa_center_credentials.py
@@ -1,30 +1,0 @@
-"""visa center credentials
-
-Revision ID: 31be58f3cea3
-Revises: 2d8fc51d711f
-Create Date: 2025-01-28 13:38:34.758271
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '31be58f3cea3'
-down_revision: Union[str, None] = '2d8fc51d711f'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
