--- alembic/versions/3a4031c89791_order_model.py
@@ -1,30 +1,0 @@
-"""order model
-
-Revision ID: 3a4031c89791
-Revises: 5a0bcb2486fb
-Create Date: 2025-02-04 15:52:07.931292
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '3a4031c89791'
-down_revision: Union[str, None] = '5a0bcb2486fb'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
