--- alembic/versions/6b784d19ad65_add_country_enum_to_visa_center_.py
@@ -1,30 +1,0 @@
-"""add country enum to visa_center_credentials
-
-Revision ID: 6b784d19ad65
-Revises: 9b66999dc910
-Create Date: 2025-01-31 00:21:16.708580
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '6b784d19ad65'
-down_revision: Union[str, None] = '9b66999dc910'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('visa_center_credentials', sa.Column('country', sa.Enum('SPAIN', name='countryenum'), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('visa_center_credentials', 'country')
-    # ### end Alembic commands ###
