--- alembic/versions/849602ad0ca5_ensures_deletion_of_related_passports.py
@@ -1,30 +1,0 @@
-"""Ensures deletion of related passports
-
-Revision ID: 849602ad0ca5
-Revises: e2109b72ce34
-Create Date: 2025-02-13 17:01:29.610015
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '849602ad0ca5'
-down_revision: Union[str, None] = 'e2109b72ce34'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
