--- alembic/versions/90240f9d7a7f_fixed_relationship_import.py
@@ -1,40 +1,0 @@
-"""Fixed relationship import
-
-Revision ID: 90240f9d7a7f
-Revises: c620e3ef6b22
-Create Date: 2025-02-12 21:17:25.483228
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '90240f9d7a7f'
-down_revision: Union[str, None] = 'c620e3ef6b22'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_index(op.f('ix_passports_credentials_id'), 'passports', ['credentials_id'], unique=False)
-    op.create_index(op.f('ix_passports_id'), 'passports', ['id'], unique=False)
-    op.create_index(op.f('ix_passports_issue_place'), 'passports', ['issue_place'], unique=False)
-    op.create_index(op.f('ix_passports_issuer_country'), 'passports', ['issuer_country'], unique=False)
-    op.create_index(op.f('ix_passports_passport_number'), 'passports', ['passport_number'], unique=True)
-    op.create_index(op.f('ix_passports_passport_type'), 'passports', ['passport_type'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_passports_passport_type'), table_name='passports')
-    op.drop_index(op.f('ix_passports_passport_number'), table_name='passports')
-    op.drop_index(op.f('ix_passports_issuer_country'), table_name='passports')
-    op.drop_index(op.f('ix_passports_issue_place'), table_name='passports')
-    op.drop_index(op.f('ix_passports_id'), table_name='passports')
-    op.drop_index(op.f('ix_passports_credentials_id'), table_name='passports')
-    # ### end Alembic commands ###
