--- alembic/versions/a39d2f4cabf7_order_model.py
@@ -1,46 +1,0 @@
-"""order model
-
-Revision ID: a39d2f4cabf7
-Revises: 3a4031c89791
-Create Date: 2025-02-04 15:57:32.636545
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'a39d2f4cabf7'
-down_revision: Union[str, None] = '3a4031c89791'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('orders',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('status', sa.Enum('PENDING', 'PROCESSING', 'COMPLETED', 'FAILED', name='orderstatus'), nullable=False),
-    sa.Column('credential_id', sa.Integer(), nullable=False),
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.ForeignKeyConstraint(['credential_id'], ['visa_center_credentials.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_orders_id'), 'orders', ['id'], unique=False)
-    op.create_index(op.f('ix_orders_status'), 'orders', ['status'], unique=False)
-    op.create_index(op.f('ix_orders_user_id'), 'orders', ['user_id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_orders_user_id'), table_name='orders')
-    op.drop_index(op.f('ix_orders_status'), table_name='orders')
-    op.drop_index(op.f('ix_orders_id'), table_name='orders')
-    op.drop_table('orders')
-    # ### end Alembic commands ###
