+++ alembic/versions/bdb93840a237_initial_commit_to_postgres.py
@@ -1,0 +1,194 @@
+"""initial commit to postgres
+
+Revision ID: bdb93840a237
+Revises: 
+Create Date: 2025-02-17 21:17:56.945227
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'bdb93840a237'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('phone_otps',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('phone_number', sa.String(), nullable=True),
+    sa.Column('otp', sa.String(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_phone_otps_id'), 'phone_otps', ['id'], unique=False)
+    op.create_index(op.f('ix_phone_otps_otp'), 'phone_otps', ['otp'], unique=False)
+    op.create_index(op.f('ix_phone_otps_phone_number'), 'phone_otps', ['phone_number'], unique=True)
+    op.create_table('users',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('username', sa.String(), nullable=True),
+    sa.Column('phone_number', sa.String(), nullable=True),
+    sa.Column('is_verified', sa.Boolean(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=True),
+    sa.Column('updated_at', sa.DateTime(), nullable=True),
+    sa.Column('first_name', sa.String(), nullable=True),
+    sa.Column('last_name', sa.String(), nullable=True),
+    sa.Column('family_name', sa.String(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_users_family_name'), 'users', ['family_name'], unique=False)
+    op.create_index(op.f('ix_users_first_name'), 'users', ['first_name'], unique=False)
+    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
+    op.create_index(op.f('ix_users_last_name'), 'users', ['last_name'], unique=False)
+    op.create_index(op.f('ix_users_phone_number'), 'users', ['phone_number'], unique=True)
+    op.create_index(op.f('ix_users_username'), 'users', ['username'], unique=False)
+    op.create_table('cards',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('card_number', sa.String(length=32), nullable=False),
+    sa.Column('expiration_date', sa.DateTime(), nullable=False),
+    sa.Column('card_holder_name', sa.String(), nullable=False),
+    sa.Column('card_type', sa.String(), nullable=False),
+    sa.Column('bank_name', sa.String(), nullable=False),
+    sa.Column('binding_id', sa.Integer(), nullable=False),
+    sa.Column('default_card', sa.Boolean(), nullable=True),
+    sa.Column('user_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_cards_bank_name'), 'cards', ['bank_name'], unique=False)
+    op.create_index(op.f('ix_cards_binding_id'), 'cards', ['binding_id'], unique=False)
+    op.create_index(op.f('ix_cards_card_holder_name'), 'cards', ['card_holder_name'], unique=False)
+    op.create_index(op.f('ix_cards_card_number'), 'cards', ['card_number'], unique=False)
+    op.create_index(op.f('ix_cards_card_type'), 'cards', ['card_type'], unique=False)
+    op.create_index(op.f('ix_cards_default_card'), 'cards', ['default_card'], unique=False)
+    op.create_index(op.f('ix_cards_expiration_date'), 'cards', ['expiration_date'], unique=False)
+    op.create_index(op.f('ix_cards_id'), 'cards', ['id'], unique=False)
+    op.create_index(op.f('ix_cards_user_id'), 'cards', ['user_id'], unique=False)
+    op.create_table('emails',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('email', sa.String(), nullable=False),
+    sa.Column('is_verified', sa.Boolean(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=True),
+    sa.Column('updated_at', sa.DateTime(), nullable=True),
+    sa.Column('user_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_emails_email'), 'emails', ['email'], unique=True)
+    op.create_index(op.f('ix_emails_id'), 'emails', ['id'], unique=False)
+    op.create_index(op.f('ix_emails_user_id'), 'emails', ['user_id'], unique=True)
+    op.create_table('tokens',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('user_id', sa.Integer(), nullable=False),
+    sa.Column('token', sa.String(), nullable=True),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_tokens_id'), 'tokens', ['id'], unique=False)
+    op.create_index(op.f('ix_tokens_token'), 'tokens', ['token'], unique=False)
+    op.create_index(op.f('ix_tokens_user_id'), 'tokens', ['user_id'], unique=False)
+    op.create_table('visa_center_credentials',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('country', sa.Enum('ES', 'GR', name='countryenum'), nullable=False),
+    sa.Column('username', sa.String(), nullable=False),
+    sa.Column('encrypted_password', sa.String(), nullable=False),
+    sa.Column('user_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_visa_center_credentials_id'), 'visa_center_credentials', ['id'], unique=False)
+    op.create_index(op.f('ix_visa_center_credentials_user_id'), 'visa_center_credentials', ['user_id'], unique=False)
+    op.create_table('orders',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('status', sa.Enum('PENDING', 'PROCESSING', 'COMPLETED', 'FAILED', name='orderstatus'), nullable=False),
+    sa.Column('credential_id', sa.Integer(), nullable=False),
+    sa.Column('user_id', sa.Integer(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=True),
+    sa.Column('updated_at', sa.DateTime(), nullable=True),
+    sa.ForeignKeyConstraint(['credential_id'], ['visa_center_credentials.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_orders_id'), 'orders', ['id'], unique=False)
+    op.create_index(op.f('ix_orders_status'), 'orders', ['status'], unique=False)
+    op.create_index(op.f('ix_orders_user_id'), 'orders', ['user_id'], unique=False)
+    op.create_table('passports',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('passport_number', sa.String(), nullable=False),
+    sa.Column('passport_type', sa.Enum('ORDINARY', 'DIPLOMATIC', 'COLLECTIVE', 'SERVICE', 'OFFICIAL', 'FOREIGNERS', 'PROTECTION', 'UN', name='passporttype'), nullable=False),
+    sa.Column('issuer_country', sa.String(), nullable=False),
+    sa.Column('issue_date', sa.DateTime(), nullable=False),
+    sa.Column('expire_date', sa.DateTime(), nullable=False),
+    sa.Column('issue_place', sa.String(), nullable=False),
+    sa.Column('name', sa.String(), nullable=False),
+    sa.Column('surname', sa.String(), nullable=False),
+    sa.Column('nationality', sa.String(), nullable=False),
+    sa.Column('user_id', sa.Integer(), nullable=False),
+    sa.Column('credentials_id', sa.Integer(), nullable=False),
+    sa.ForeignKeyConstraint(['credentials_id'], ['visa_center_credentials.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_passports_credentials_id'), 'passports', ['credentials_id'], unique=False)
+    op.create_index(op.f('ix_passports_id'), 'passports', ['id'], unique=False)
+    op.create_index(op.f('ix_passports_issue_place'), 'passports', ['issue_place'], unique=False)
+    op.create_index(op.f('ix_passports_issuer_country'), 'passports', ['issuer_country'], unique=False)
+    op.create_index(op.f('ix_passports_passport_number'), 'passports', ['passport_number'], unique=False)
+    op.create_index(op.f('ix_passports_passport_type'), 'passports', ['passport_type'], unique=False)
+    op.create_index(op.f('ix_passports_user_id'), 'passports', ['user_id'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_passports_user_id'), table_name='passports')
+    op.drop_index(op.f('ix_passports_passport_type'), table_name='passports')
+    op.drop_index(op.f('ix_passports_passport_number'), table_name='passports')
+    op.drop_index(op.f('ix_passports_issuer_country'), table_name='passports')
+    op.drop_index(op.f('ix_passports_issue_place'), table_name='passports')
+    op.drop_index(op.f('ix_passports_id'), table_name='passports')
+    op.drop_index(op.f('ix_passports_credentials_id'), table_name='passports')
+    op.drop_table('passports')
+    op.drop_index(op.f('ix_orders_user_id'), table_name='orders')
+    op.drop_index(op.f('ix_orders_status'), table_name='orders')
+    op.drop_index(op.f('ix_orders_id'), table_name='orders')
+    op.drop_table('orders')
+    op.drop_index(op.f('ix_visa_center_credentials_user_id'), table_name='visa_center_credentials')
+    op.drop_index(op.f('ix_visa_center_credentials_id'), table_name='visa_center_credentials')
+    op.drop_table('visa_center_credentials')
+    op.drop_index(op.f('ix_tokens_user_id'), table_name='tokens')
+    op.drop_index(op.f('ix_tokens_token'), table_name='tokens')
+    op.drop_index(op.f('ix_tokens_id'), table_name='tokens')
+    op.drop_table('tokens')
+    op.drop_index(op.f('ix_emails_user_id'), table_name='emails')
+    op.drop_index(op.f('ix_emails_id'), table_name='emails')
+    op.drop_index(op.f('ix_emails_email'), table_name='emails')
+    op.drop_table('emails')
+    op.drop_index(op.f('ix_cards_user_id'), table_name='cards')
+    op.drop_index(op.f('ix_cards_id'), table_name='cards')
+    op.drop_index(op.f('ix_cards_expiration_date'), table_name='cards')
+    op.drop_index(op.f('ix_cards_default_card'), table_name='cards')
+    op.drop_index(op.f('ix_cards_card_type'), table_name='cards')
+    op.drop_index(op.f('ix_cards_card_number'), table_name='cards')
+    op.drop_index(op.f('ix_cards_card_holder_name'), table_name='cards')
+    op.drop_index(op.f('ix_cards_binding_id'), table_name='cards')
+    op.drop_index(op.f('ix_cards_bank_name'), table_name='cards')
+    op.drop_table('cards')
+    op.drop_index(op.f('ix_users_username'), table_name='users')
+    op.drop_index(op.f('ix_users_phone_number'), table_name='users')
+    op.drop_index(op.f('ix_users_last_name'), table_name='users')
+    op.drop_index(op.f('ix_users_id'), table_name='users')
+    op.drop_index(op.f('ix_users_first_name'), table_name='users')
+    op.drop_index(op.f('ix_users_family_name'), table_name='users')
+    op.drop_table('users')
+    op.drop_index(op.f('ix_phone_otps_phone_number'), table_name='phone_otps')
+    op.drop_index(op.f('ix_phone_otps_otp'), table_name='phone_otps')
+    op.drop_index(op.f('ix_phone_otps_id'), table_name='phone_otps')
+    op.drop_table('phone_otps')
+    # ### end Alembic commands ###
