--- alembic/versions/d02e5e065bd5_passports_model_and_relationship_with_.py
@@ -1,56 +1,0 @@
-"""passports model and relationship with visa credentials
-
-Revision ID: d02e5e065bd5
-Revises: e88fd8c064a1
-Create Date: 2025-02-12 19:12:22.957078
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'd02e5e065bd5'
-down_revision: Union[str, None] = 'e88fd8c064a1'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('passports',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('passport_number', sa.String(), nullable=False),
-    sa.Column('passport_type', sa.Enum('ORDINARY', 'DIPLOMATIC', 'COLLECTIVE', 'SERVICE', 'OFFICIAL', 'FOREIGNERS', 'PROTECTION', 'UN', name='passporttype'), nullable=False),
-    sa.Column('issuer_country', sa.String(), nullable=False),
-    sa.Column('issue_date', sa.DateTime(), nullable=False),
-    sa.Column('expire_date', sa.DateTime(), nullable=False),
-    sa.Column('issue_place', sa.String(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('surname', sa.String(), nullable=False),
-    sa.Column('nationality', sa.String(), nullable=False),
-    sa.Column('credentials_is', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['credentials_is'], ['visa_center_credentials.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_passports_credentials_is'), 'passports', ['credentials_is'], unique=False)
-    op.create_index(op.f('ix_passports_id'), 'passports', ['id'], unique=False)
-    op.create_index(op.f('ix_passports_issue_place'), 'passports', ['issue_place'], unique=False)
-    op.create_index(op.f('ix_passports_issuer_country'), 'passports', ['issuer_country'], unique=False)
-    op.create_index(op.f('ix_passports_passport_number'), 'passports', ['passport_number'], unique=True)
-    op.create_index(op.f('ix_passports_passport_type'), 'passports', ['passport_type'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_passports_passport_type'), table_name='passports')
-    op.drop_index(op.f('ix_passports_passport_number'), table_name='passports')
-    op.drop_index(op.f('ix_passports_issuer_country'), table_name='passports')
-    op.drop_index(op.f('ix_passports_issue_place'), table_name='passports')
-    op.drop_index(op.f('ix_passports_id'), table_name='passports')
-    op.drop_index(op.f('ix_passports_credentials_is'), table_name='passports')
-    op.drop_table('passports')
-    # ### end Alembic commands ###
