--- alembic/versions/e2109b72ce34_fixed_relationship_consistence.py
@@ -1,30 +1,0 @@
-"""Fixed relationship consistence
-
-Revision ID: e2109b72ce34
-Revises: 90240f9d7a7f
-Create Date: 2025-02-12 21:27:44.327547
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'e2109b72ce34'
-down_revision: Union[str, None] = '90240f9d7a7f'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
