--- alembic/versions/e88fd8c064a1_visa_credentials_issue_fixed_in_order_.py
@@ -1,30 +1,0 @@
-"""visa_credentials issue fixed in order model
-
-Revision ID: e88fd8c064a1
-Revises: a39d2f4cabf7
-Create Date: 2025-02-04 21:32:19.450261
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'e88fd8c064a1'
-down_revision: Union[str, None] = 'a39d2f4cabf7'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
