+++ docker-compose.yml
@@ -1,14 +1,19 @@
 version: '3.8'
+
 services:
+  # Zookeeper Service
   zookeeper:
     image: confluentinc/cp-zookeeper:latest
+    container_name: neovisa_zookeeper
     environment:
       ZOOKEEPER_CLIENT_PORT: 2181
     ports:
       - "2181:2181"
 
+  # Kafka Service
   kafka:
     image: confluentinc/cp-kafka:latest
+    container_name: neovisa_kafka
     environment:
       KAFKA_OFFSETS_TOPIC_REPLICATION_FACTOR: 1
       KAFKA_ADVERTISED_LISTENERS: INSIDE://kafka:9093,PLAINTEXT://localhost:29092
@@ -21,3 +26,23 @@ services:
       - "9093:9093"
     depends_on:
       - zookeeper
+
+  # PostgreSQL Service
+  postgres:
+    image: postgres:latest
+    container_name: neovisa_postgres
+    restart: always
+    ports:
+      - "${POSTGRES_PORT}:5432"
+    environment:
+      POSTGRES_USER: ${POSTGRES_USER}
+      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
+      POSTGRES_DB: ${POSTGRES_DB}
+      POSTGRES_PORT: 5432
+    volumes:
+      - postgres_data:/var/lib/postgresql/data
+
+
+
+volumes:
+  postgres_data:
