+++ user/models.py
@@ -24,6 +24,7 @@ class User(Base):
     visa_credentials = Relationship('VisaCenterCredentials', back_populates='user', passive_deletes=True)
     orders = Relationship('Order', back_populates='user', passive_deletes=True)
     passports = Relationship('Passport', back_populates='user', cascade='all, delete-orphan')
+    cards = Relationship('Card', back_populates='user',  cascade='all, delete-orphan')
 
 
 class Email(Base):
